requirements.txt
Flask
line-bot-sdk
gunicorn
openai